backend/requirements.txt
flask
flask-cors
pillow
tqdm
torch
transformers
scikit-learn
